+++ .pre-commit-config.yaml
@@ -1,5 +1,5 @@
 repos:
   - repo: https://github.com/golangci/golangci-lint
-    rev: v1.42.1
+    rev: v1.59.1
     hooks:
       - id: golangci-lint
