+++ assets/embed.go
@@ -1,4 +1,4 @@
-package ui
+package assets
 
 import "embed"
