+++ assets/html/analytics.page.tmpl
