+++ assets/html/links.partial.tmpl
