+++ assets/html/shorten.partial.tmpl
