+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/danielmichaels/shortlink-go
 
-go 1.17
+go 1.22
 
 require (
 	github.com/go-chi/chi/v5 v5.0.7
